split up auth code and add oauth2 reporting tests
